Add: Flask API for Article Persistance
Added a Flask API with a couple of endpoints. Currently it does a CRON job every few hours, where it runs our parallel scraping cpp script, and persists the returned json files into mongoDB. Created setup.py to make a platform-agnostic setup script that builds the cpp script into an executable that can be used by flask. Also had to create a CMake for the cpp script to control the build.